1.changed jpa repository type parameters from int to long 2. implemented add game

+++ Backend/src/main/java/com/example/frogcrew/game/repository/GameScheduleRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GameScheduleRepository extends JpaRepository<GameSchedule, Integer> {
+public interface GameScheduleRepository extends JpaRepository<GameSchedule, Long> {
 }
